Change to grid settings for color/opacity
These changes implement color/opacity settings for the Grid and Pixel grid
within the Tool Configuration dialog.

Initial changes to support actually using that opacity to draw the grid
are in place, but more work is required to coax the proper functionality
out of allegro, which I have no clue at this time how to do.

Included is an updated widget_loader for specifying initial color for the
colorpicker widget.  It turned out not to be strictly necessary for these
changes but could come in handy in the future and really doesn't require a
large amount of code to handle.