Experimental color slider index quantization
This change introduces an experimental feature which, once complete, would
allow users to choose a "Quantize to Index" option to turn this
functionality on or off.

The change itself is very basic, if quantization is turned on then pass
the rgb values to the Palette::findBestFit() method after calculating the
nominal color value normally.

As currently implemented this quantization is always on, even in the
palette editor, which should, in practice, not be quantized at all.

I stopped integrating the change into the color pickers in order to get
some feedback before comitting more time.